requirements.txt
-i https://pypi.org/simple
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
django-admin-sortable==2.2.3
django-allauth==0.41.0
django-better-admin-arrayfield==1.0.5
django-crispy-forms==1.9.0
django-debug-toolbar==2.2
django-environ==0.4.5
django-extensions==2.2.8
django==3.0.3
idna==2.9
oauthlib==3.1.0
psycopg2-binary==2.8.4
python3-openid==3.1.0
pytz==2019.3
requests-oauthlib==1.3.0
requests==2.23.0
sentry-sdk==0.14.1
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
werkzeug==1.0.0
